Fix partial issues
