+++ src/pages/Signup/Signup.css
@@ -6,7 +6,7 @@
       rgba(0, 0, 0, 1),
       rgba(0, 0, 0, 0)
     ),
-    url("../../images/login-background.jpg");
+    url("../../assets/images/login-background.jpg");
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
