+++ src/pages/Signup/Signup.tsx
@@ -11,10 +11,8 @@ import {
   Typography,
 } from "@mui/material";
 import React from "react";
-import {
-  FacebookIcon,
-  GoogleIcon,
-} from "../../components/core/Icon/CustomIcons";
+import { FacebookIcon } from "../../assets/icons/facebook";
+import { GoogleIcon } from "../../assets/icons/google";
 import "./Signup.css";
 
 export const Signup = () => {
